requirements.txt
astroid==2.14.2
attrs==22.2.0
black==23.1.0
#bzip2==1.0.8
#ca-certificates==2023.01.10
cachetools==5.3.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
coverage==7.2.1
dill==0.3.6
discord.py==2.2.2
distlib==0.3.6
duckdb==0.7.1
exceptiongroup==1.1.0
fastapi==0.92.0
filelock==3.9.0
flit-core==3.6.0
idna==3.4
iniconfig==2.0.0
isort==5.9.3
langchain==0.0.100
lazy-object-proxy==1.6.0
#libffi==3.4.2
mccabe==0.7.0
mypy==1.0.1
#ncurses==6.4
numpy==1.24.2
#openssl==1.1.1t
openai==0.27.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
pip==23.0.1
platformdirs==3.0.0
pluggy==1.0.0
ptvsd==4.3.2
pyarrow==11.0.0
pydantic==1.10.5
PyGithub==1.58.1
pylance==0.3.10
pylint==2.16.2
pyproject-api==1.5.0
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pyyaml==6.0
pyyaml-include==1.3
#readline==8.2
requests==2.28.2
setuptools==65.6.3
six==1.16.0
#sqlite==3.40.1
#tk==8.6.12
types-PyYaml==6.0.12.8
types-requests==2.28.11.15
tomli==2.0.1
tomlkit==0.11.1
tox==4.4.6
typing-extensions==4.5.0
#tzdata==2022g
urllib3==1.26.14
virtualenv==20.20.0
wheel==0.38.4
wrapt==1.14.1
#xz==5.2.10
#zlib==1.2.13
